gcd works!!!
It turns out that a lot was hanging on installInstructions. First, because assemble produces two distinct instruction lists, the instructions needed to be installed on both of those lists separately. Second, the operation functions (especially rem)  weren't looking up the right register values (or something like that). This was fixed by moving installInstructions into execute, thereby "reinstalling" the instructions every go-around. This is a grossly inefficient solution, but it works, and at the moment that's a big victory.

The stack and its operations (save and restore) haven't been tested yet. Hopefully that will go smoothly.